PP-4314 Split the AuthorisationGatewayRequest between Card and Apple Pay
